BUG-88191: Follow on cleanup of Kerberos related changes.

+++ main/src/main/java/com/hortonworks/beacon/api/AbstractResourceManager.java
@@ -67,6 +67,7 @@ import com.hortonworks.beacon.store.BeaconStoreService;
 import com.hortonworks.beacon.store.bean.PolicyInstanceBean;
 import com.hortonworks.beacon.util.ClusterStatus;
 import org.apache.commons.lang3.StringUtils;
+import org.apache.hadoop.security.UserGroupInformation;
 import org.codehaus.jackson.map.ObjectMapper;
 
 import javax.persistence.EntityManager;
@@ -467,8 +468,10 @@ public abstract class AbstractResourceManager {
         String remoteEndPoint = PolicyHelper.getRemoteBeaconEndpoint(policy);
         String remoteClusterName = PolicyHelper.getRemoteClusterName(policy);
         try {
-            BeaconClient remoteClient = new BeaconClient(remoteEndPoint);
-            remoteClient.deletePolicy(policy.getName(), true);
+            if (!UserGroupInformation.isSecurityEnabled()) {
+                BeaconClient remoteClient = new BeaconClient(remoteEndPoint);
+                remoteClient.deletePolicy(policy.getName(), true);
+            }
         } catch (BeaconClientException e) {
             throw BeaconWebException.newAPIException(MessageCode.MAIN_000025.name(),
                     Response.Status.fromStatusCode(e.getStatus()), e, remoteClusterName, e.getMessage());
@@ -685,7 +688,9 @@ public abstract class AbstractResourceManager {
             // No policy status sync needed for HCFS
             return;
         }
-
+        if (UserGroupInformation.isSecurityEnabled()) {
+            return;
+        }
         String remoteBeaconEndpoint = PolicyHelper.getRemoteBeaconEndpoint(policy);
         try {
             //TODO Check is there any sync status job scheduled. removed them and update it.
@@ -863,10 +868,12 @@ public abstract class AbstractResourceManager {
     private void syncPolicyInRemote(ReplicationPolicy policy, String policyName,
                                     String remoteBeaconEndpoint, String remoteClusterName) {
         try {
-            BeaconClient remoteClient = new BeaconClient(remoteBeaconEndpoint);
-            remoteClient.syncPolicy(policyName, policy.toString());
-            BeaconEvents.createEvents(Events.SYNCED, EventEntityType.POLICY,
-                    PersistenceHelper.getPolicyBean(policy), getEventInfo(policy, false));
+            if (!UserGroupInformation.isSecurityEnabled()) {
+                BeaconClient remoteClient = new BeaconClient(remoteBeaconEndpoint);
+                remoteClient.syncPolicy(policyName, policy.toString());
+                BeaconEvents.createEvents(Events.SYNCED, EventEntityType.POLICY,
+                        PersistenceHelper.getPolicyBean(policy), getEventInfo(policy, false));
+            }
         } catch (BeaconClientException e) {
             throw BeaconWebException.newAPIException(MessageCode.MAIN_000025.name(),
                     Response.Status.fromStatusCode(e.getStatus()), e, remoteClusterName, e.getMessage());
